feat 입력된 금액이 몇장의 로또를 살 수 있는지
